requirements.txt
fastapi
uvicorn
python-docx
python-multipart
pydantic
jinja2
